web/imagehost/src/requirements.txt
starlette==0.36.3
itsdangerous==2.1.2
jinja2==3.1.3
aiomysql==0.2.0
python-multipart==0.0.9
pyjwt==2.8.0
cryptography==42.0.5
Pillow==10.3.0